requirements.txt
Django==6.0.4
django-ninja==1.6.2
django-unfold>=0.90.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.9
Pillow==10.4.0
pydantic==2.12.5
python-multipart==0.0.20
email-validator==2.2.0
gunicorn==22.0.0
uvicorn==0.34.0
pytest==8.3.4
pytest-django==4.9.0
